Added address masks
git-svn-id: http://avr-xboot.googlecode.com/svn/trunk/xboot@8 08c6285c-dfd1-04b1-8611-c7c1110d9cf2

+++ xboot.h
@@ -158,7 +158,7 @@
 #endif
 #endif
 
-typedef unsigned long ADDR_T;
+typedef uint32_t ADDR_T;
 
 // Functions
 unsigned char __attribute__ ((noinline)) ow_slave_read_bit(void);
